Parallel iterators over ImageBuffers
Implments rayon `ParallelIterator` for iterators over the information on an `ImageBuffer` through new methods: par_rows, par_pixels, par_rows_mut and par_pixels_mut.

`par_rows(_mut)` returns `ParRows(Mut)`, a  parallel iterator over all the rows in an `ImageBuffer` that returns a `Row(Mut)` as its Item.

A `Row(Mut)` references the pixels in the corresponding row of the original `ImageBuffer`,  has some methods to access it's  pixels and implements the `GenericImage` and `GenericImageView` traits

the `par_pixels(_mut)` method returns a `Pixels(Mut)` iterator that iterates over all the pixels in the image in parallel returning a (mutable) reference to each `Pixel` in the `ImageBuffer`